requirements.txt
# Dependencias del proyecto Python
Flask
requests
python-dotenv
BASE_URL = "https://api.vapi.ai" # <-- LÍNEA TEMPORAL PARA DEPURAR
